+++ packages/pi/src/loader/index.ts
@@ -1,0 +1,2 @@
+export type { ProviderConfigInput, ResolvedRequestAuth } from "#pi/loader/model-registry";
+export { clearApiKeyCache, ModelRegistry } from "#pi/loader/model-registry";